+++ db/db.go
@@ -12,3 +12,10 @@ type AuditRecord struct {
 	Status     string
 	Card       string
 }
+
+type LogRecord struct {
+	Timestamp  time.Time
+	Operation  string
+	Controller uint32
+	Detail     string
+}