+++ scripts/sqlite3.sql
@@ -23,6 +23,13 @@ CREATE TABLE Audit (
     Card       TEXT     DEFAULT ''
 );
 
+CREATE TABLE OperationsLog (
+    Timestamp  DATETIME DEFAULT (datetime(CURRENT_TIMESTAMP, 'localtime')),
+    Operation  TEXT     DEFAULT '',
+    Controller INTEGER  NULL,
+    Detail     TEXT     DEFAULT ''
+);
+
 INSERT INTO ACL    (Name, CardNumber,PIN,StartDate,EndDate,GreatHall,Gryffindor,HufflePuff,Ravenclaw,Slytherin,Kitchen,Dungeon,Hogsmeade)
             VALUES ('Albus Dumbledore', 10058400, 0, '2023-01-01', '2023-12-31', 1,1,1,1,1,1,1,1);